# documentation/swagger-api.yaml
swagger: "2.0"
info:
  description: "EPRO Projekt WS2021"
  version: "1.0.0"
  title: "EPRO OKR Projekt"
host: "localhost:8080"
basePath: "/api/v1"
schemes:
  - "http"
paths:
  /companyObjectives:
    get:
      tags:
        - "Company Objectives"
      summary: "Get all Companyobjectives"
      description: "Gets all companyobjectives"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CompanyObjective"
    post:
      tags:
        - "Company Objectives"
      summary: "Add a new CompanyObjective"
      description: ""
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Business Unit Objective to be added"
          required: true
          schema:
            $ref: "#/definitions/BusinessUnitObjective"
      responses:
        "405":
          description: "Invalid input"
  /companyObjectives/{id}:
    get:
      tags:
        - "Company Objectives"
      summary: "Get a CompanyObjective"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of CO to return"
          required: true
          type: "integer"
      responses:
        200:
          description: A companyobjective object
          schema:
            $ref: "#/definitions/CompanyObjective"
        400:
          description: Bad request. CO ID must be an integer and bigger than 0.
        404:
          description: CO with the specified ID not found
    put:
      tags:
        - "Company Objectives"
      summary: "Update a CompanyObjective"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Company Objective to update"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Business Unit Objective to be added"
          required: true
          schema:
            $ref: "#/definitions/BusinessUnitObjective"
      responses:
        200:
          description: "Successfully Updated the Company Objective"

    delete:
      tags:
        - "Company Objectives"
      summary: "Deletes a CompanyObjective"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Company Objective to be deleted"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully deleted company objective"
        404:
          description: "Did not find Company Objective"
...
  /companyobjective/{coId}/keyResults:
    get:
      tags:
        - "Company Key Results"
      summary: "Returns all Key Results of the specified Company Objective"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "coId"
          description: "The CompanyObjective for which the Keyresult are to be returned"
          required: true
          type: "integer"
      responses:
        200:
          description: "successfully got all Key Results"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CompanyKeyResult"
    post:
      tags:
        - "Company Key Results"
      summary: "Create a new Key Result for the specified Company Objective"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "coId"
          description: "The CompanyObjective for which the new Keyresult shall be created"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "Key Result to be added"
          required: true
          schema:
            $ref: "#/definitions/CompanyKeyResult"
      responses:
        200:
          description: "successfully created key result"
  /companyKeyResults/{id}:
    get:
      tags:
        - "Company Key Results"
      summary: "Return the specified Company Key Result"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "The id of the C-KR to be returned"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully returned keyresult"
          schema:
            $ref: "#/definitions/CompanyKeyResult"
    put:
      tags:
        - "Company Key Results"
      summary: "Update the specified Company Key Result"
      consumes:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "The id of the C-KR to be updated"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "The updated Key Result Object"
          required: true
          schema:
            $ref: "#/definitions/CompanyKeyResult"
      responses:
        200:
          description: "Successfully updated"

    delete:
      tags:
        - "Company Key Results"
      summary: "Delete the specified Company Key Result"
      parameters:
        - in: path
          name: id
          description: The id of the C-KR to be deleted
          required: true
          type: integer
      responses:
        200:
          description: Successfully Deleted
        404:
          description: Not found
  /companyKeyResults/{id}/history:
    get:
      tags:
        - "Company Key Results"
      summary: "Returns the history of the specified Company Key Result"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "The id of the C-KR of which the history is to be returned"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully returned CompanyKeyResult History"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CompanyKeyResult"
...
  /user/{id}:
    get:
      tags:
        - "Users"
      summary: "Get a User"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of User to return"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully returned user object"
          schema:
            $ref: "#/definitions/User"
    put:
      tags:
        - "Users"
      summary: "Update a User"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the User to update"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "The updated user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "Successfully updated the User"
    delete:
      tags:
        - "Users"
      summary: "Deletes a User"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the User to be deleted"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully deleted the User"
        404:
          description: "Did not find User"
...
definitions:
  CompanyObjective:
    type: "object"
    properties:
      id:
        type: "integer"
      achievement:
        type: "number"
      name:
        type: "string"
...
  BusinessUnitObjective:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      achievement:
        type: "number"
      businessUnitId:
        type: "integer"
  CompanyKeyResult:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      confidenceLvl:
        type: "number"
      comment:
        type: "string"
      goalValue:
        type: "number"
      currentValue:
        type: "number"
      achievement:
        type: "number"
      companyObjectiveId:
        type: "integer"
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      surname:
        type: "string"
      password:
        type: "string"
      roleId:
        type: "integer"
      businessUnitId:
        type: "integer"
...